+++ changelog.md
@@ -28,6 +28,7 @@ Uiua is not yet stable.
 - Add the `uiua build` command, which emits a `.uasm` bytecode file
 - `uiua run` can now run a `.uasm` bytecode file
 - `uiua stand` now embeds the bytecode assembly in the executable
+- Multiple compiler errors can now be emitted at once
 - Bug and crash fixes
 - Performance improvements
 ### Website